test(types): improve naming, folder structure, test all of the utilities
