+++ src/__tests__/combinators/sequence.spec-d.ts
